+++ plugins/repository-azure/src/internalClusterTest/java/org/opensearch/repositories/azure/AzureStorageCleanupThirdPartyTests.java
@@ -32,6 +32,7 @@
 
 package org.opensearch.repositories.azure;
 
+import org.opensearch.core.common.Strings;
 import reactor.core.scheduler.Schedulers;
 
 import com.azure.core.util.Context;
@@ -43,7 +44,6 @@ import org.junit.AfterClass;
 import org.opensearch.action.ActionRunnable;
 import org.opensearch.action.support.PlainActionFuture;
 import org.opensearch.action.support.master.AcknowledgedResponse;
-import org.opensearch.common.Strings;
 import org.opensearch.common.collect.Tuple;
 import org.opensearch.common.settings.MockSecureSettings;
 import org.opensearch.common.settings.SecureSettings;