+++ server/src/main/java/org/opensearch/action/admin/indices/close/CloseIndexResponse.java
@@ -32,7 +32,7 @@
 package org.opensearch.action.admin.indices.close;
 
 import org.opensearch.LegacyESVersion;
-import org.opensearch.OpenSearchException;
+import org.opensearch.BaseOpenSearchException;
 import org.opensearch.action.support.DefaultShardOperationFailedException;
 import org.opensearch.action.support.master.ShardsAcknowledgedResponse;
 import org.opensearch.common.Nullable;
@@ -182,7 +182,7 @@ public class CloseIndexResponse extends ShardsAcknowledgedResponse {
                     builder.field("closed", false);
                     if (exception != null) {
                         builder.startObject("exception");
-                        OpenSearchException.generateFailureXContent(builder, params, exception, true);
+                        BaseOpenSearchException.generateFailureXContent(builder, params, exception, true);
                         builder.endObject();
                     } else {
                         builder.startObject("failedShards");