+++ server/src/main/java/org/opensearch/action/admin/indices/segments/PitSegmentsRequest.java
@@ -10,9 +10,9 @@ package org.opensearch.action.admin.indices.segments;
 
 import org.opensearch.action.ActionRequestValidationException;
 import org.opensearch.action.support.broadcast.BroadcastRequest;
-import org.opensearch.common.Strings;
 import org.opensearch.common.io.stream.StreamInput;
 import org.opensearch.common.io.stream.StreamOutput;
+import org.opensearch.core.common.Strings;
 import org.opensearch.core.xcontent.XContentParser;
 
 import java.io.IOException;