+++ server/src/main/java/org/opensearch/action/search/AbstractSearchAsyncAction.java
@@ -34,8 +34,8 @@ package org.opensearch.action.search;
 
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.message.ParameterizedMessage;
+import org.opensearch.BaseOpenSearchException;
 import org.opensearch.ExceptionsHelper;
-import org.opensearch.OpenSearchException;
 import org.opensearch.Version;
 import org.opensearch.action.ActionListener;
 import org.opensearch.action.NoShardAvailableActionException;
@@ -367,7 +367,7 @@ abstract class AbstractSearchAsyncAction<Result extends SearchPhaseResult> exten
             final ShardOperationFailedException[] shardSearchFailures = ExceptionsHelper.groupBy(buildShardFailures());
             Throwable cause = shardSearchFailures.length == 0
                 ? null
-                : OpenSearchException.guessRootCauses(shardSearchFailures[0].getCause())[0];
+                : BaseOpenSearchException.guessRootCauses(shardSearchFailures[0].getCause())[0];
             logger.debug(() -> new ParameterizedMessage("All shards failed for phase: [{}]", getName()), cause);
             onPhaseFailure(currentPhase, "all shards failed", cause);
         } else {
@@ -381,7 +381,7 @@ abstract class AbstractSearchAsyncAction<Result extends SearchPhaseResult> exten
                     if (logger.isDebugEnabled()) {
                         int numShardFailures = shardSearchFailures.length;
                         shardSearchFailures = ExceptionsHelper.groupBy(shardSearchFailures);
-                        Throwable cause = OpenSearchException.guessRootCauses(shardSearchFailures[0].getCause())[0];
+                        Throwable cause = BaseOpenSearchException.guessRootCauses(shardSearchFailures[0].getCause())[0];
                         logger.debug(
                             () -> new ParameterizedMessage("{} shards failed for phase: [{}]", numShardFailures, getName()),
                             cause