+++ server/src/main/java/org/opensearch/action/support/DefaultShardOperationFailedException.java
@@ -32,6 +32,7 @@
 
 package org.opensearch.action.support;
 
+import org.opensearch.BaseExceptionsHelper;
 import org.opensearch.OpenSearchException;
 import org.opensearch.ExceptionsHelper;
 import org.opensearch.action.ShardOperationFailedException;
@@ -135,7 +136,7 @@ public class DefaultShardOperationFailedException extends ShardOperationFailedEx
         builder.field("status", status.name());
         if (reason != null) {
             builder.startObject("reason");
-            OpenSearchException.generateThrowableXContent(builder, params, cause);
+            BaseExceptionsHelper.generateThrowableXContent(builder, params, cause);
             builder.endObject();
         }
         return builder;