+++ server/src/main/java/org/opensearch/rest/action/search/RestExplainAction.java
@@ -34,7 +34,7 @@ package org.opensearch.rest.action.search;
 
 import org.opensearch.action.explain.ExplainRequest;
 import org.opensearch.client.node.NodeClient;
-import org.opensearch.common.Strings;
+import org.opensearch.core.common.Strings;
 import org.opensearch.index.query.QueryBuilder;
 import org.opensearch.rest.BaseRestHandler;
 import org.opensearch.rest.RestRequest;