+++ test/framework/src/main/java/org/opensearch/transport/AbstractSimpleTransportTestCase.java
@@ -39,6 +39,7 @@ import org.apache.logging.log4j.message.ParameterizedMessage;
 import org.apache.logging.log4j.util.Supplier;
 import org.apache.lucene.util.CollectionUtil;
 import org.apache.lucene.util.Constants;
+import org.opensearch.BaseExceptionsHelper;
 import org.opensearch.LegacyESVersion;
 import org.opensearch.OpenSearchException;
 import org.opensearch.ExceptionsHelper;
@@ -558,7 +559,7 @@ public abstract class AbstractSimpleTransportTestCase extends OpenSearchTestCase
             serviceA.submitRequest(nodeB, ACTION, TransportRequest.Empty.INSTANCE, EmptyTransportResponseHandler.INSTANCE_SAME).get();
         } catch (ExecutionException e) {
             assertThat(e.getCause(), instanceOf(OpenSearchException.class));
-            assertThat(ExceptionsHelper.unwrapCause(e.getCause()).getMessage(), equalTo("simulated"));
+            assertThat(BaseExceptionsHelper.unwrapCause(e.getCause()).getMessage(), equalTo("simulated"));
         }
 
         // use assert busy as callbacks are called on a different thread
@@ -577,7 +578,7 @@ public abstract class AbstractSimpleTransportTestCase extends OpenSearchTestCase
             serviceB.submitRequest(nodeA, ACTION, TransportRequest.Empty.INSTANCE, EmptyTransportResponseHandler.INSTANCE_SAME).get();
         } catch (ExecutionException e) {
             assertThat(e.getCause(), instanceOf(OpenSearchException.class));
-            assertThat(ExceptionsHelper.unwrapCause(e.getCause()).getMessage(), equalTo("simulated"));
+            assertThat(BaseExceptionsHelper.unwrapCause(e.getCause()).getMessage(), equalTo("simulated"));
         }
 
         // use assert busy as callbacks are called on a different thread
@@ -597,7 +598,7 @@ public abstract class AbstractSimpleTransportTestCase extends OpenSearchTestCase
             serviceA.submitRequest(nodeA, ACTION, TransportRequest.Empty.INSTANCE, EmptyTransportResponseHandler.INSTANCE_SAME).get();
         } catch (ExecutionException e) {
             assertThat(e.getCause(), instanceOf(OpenSearchException.class));
-            assertThat(ExceptionsHelper.unwrapCause(e.getCause()).getMessage(), equalTo("simulated"));
+            assertThat(BaseExceptionsHelper.unwrapCause(e.getCause()).getMessage(), equalTo("simulated"));
         }
 
         // use assert busy as callbacks are called on a different thread
@@ -1651,7 +1652,7 @@ public abstract class AbstractSimpleTransportTestCase extends OpenSearchTestCase
 
                 @Override
                 public void handleException(TransportException exp) {
-                    Throwable cause = ExceptionsHelper.unwrapCause(exp);
+                    Throwable cause = BaseExceptionsHelper.unwrapCause(exp);
                     assertThat(cause, instanceOf(ConnectTransportException.class));
                     assertThat(((ConnectTransportException) cause).node(), equalTo(nodeA));
                 }
@@ -1662,7 +1663,7 @@ public abstract class AbstractSimpleTransportTestCase extends OpenSearchTestCase
             res.txGet();
             fail("exception should be thrown");
         } catch (Exception e) {
-            Throwable cause = ExceptionsHelper.unwrapCause(e);
+            Throwable cause = BaseExceptionsHelper.unwrapCause(e);
             assertThat(cause, instanceOf(ConnectTransportException.class));
             assertThat(((ConnectTransportException) cause).node(), equalTo(nodeA));
         }